feat(KOSync): round-trip test the XPath anchors, and fix what it caught
ParagraphStreamer moves to lib/KOReaderSync/VisibleTextStreamer.h so the
counting side can be tested on the host: ProgressMapper.cpp itself pulls in
Epub, Section and the renderer, none of which a counting test needs. The new
suite asserts the property that matters -- an anchor produced from a visible
offset resolves back to that same offset -- across nested markup, comments,
processing instructions, CDATA, hidden and pagebreak subtrees, DOCTYPE,
self-closing tags and non-visible elements.

Three of those failed, and are fixed here:

- comments and processing instructions were read as ordinary tags, so the
  resolver's text()[N] anchors after them could not be matched at all
- CDATA content was skipped entirely by this scanner while the resolver counts
  it as visible text
- a markup boundary is now a text-node boundary, applied lazily on the next
  visible codepoint (the resolver's rule) and cleared by any intervening tag
