+++ test/CMakeLists.txt
@@ -53,6 +53,7 @@ add_subdirectory(reading_stats)
 add_subdirectory(xml_parser_utils)
 add_subdirectory(token_boundary)
 add_subdirectory(koreader_xpath_resolver)
+add_subdirectory(koreader_roundtrip)
 add_subdirectory(progress_comparison)
 add_subdirectory(font_cache_manager)
 add_subdirectory(css_parser)
