+++ test/koreader_roundtrip/CMakeLists.txt
@@ -1,0 +1,30 @@
+add_executable(KOReaderRoundTripTest
+  KOReaderRoundTripTest.cpp
+  ${REPO_ROOT}/lib/KOReaderSync/ChapterXPathResolver.cpp
+  ${REPO_ROOT}/lib/Utf8/Utf8.cpp
+  ${REPO_ROOT}/lib/Epub/Epub/htmlEntities.cpp
+  ${REPO_ROOT}/lib/expat/xmlparse.c
+  ${REPO_ROOT}/lib/expat/xmlrole.c
+  ${REPO_ROOT}/lib/expat/xmltok.c
+)
+
+target_include_directories(KOReaderRoundTripTest PRIVATE
+  ${CMAKE_CURRENT_SOURCE_DIR}/stubs
+  ${REPO_ROOT}/lib/KOReaderSync
+  ${REPO_ROOT}/lib/Epub
+  ${REPO_ROOT}/lib/Utf8
+  ${REPO_ROOT}/lib/XmlParserUtils
+  ${REPO_ROOT}/lib/expat
+)
+
+target_compile_definitions(KOReaderRoundTripTest PRIVATE
+  XML_CONTEXT_BYTES=1024
+  XML_GE=0
+)
+
+target_link_libraries(KOReaderRoundTripTest PRIVATE
+  crosspoint_test_common
+  GTest::gtest_main
+)
+
+gtest_discover_tests(KOReaderRoundTripTest)
