+++ test/koreader_roundtrip/stubs/Logging.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define LOG_DBG(...)
+#define LOG_ERR(...)
