+++ test/koreader_roundtrip/stubs/Print.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <cstddef>
+#include <cstdint>
+
+class Print {
+ public:
+  virtual ~Print() = default;
+  virtual size_t write(uint8_t value) = 0;
+  virtual size_t write(const uint8_t* buffer, size_t size) {
+    size_t written = 0;
+    while (written < size) {
+      written += write(buffer[written]);
+    }
+    return written;
+  }
+};
